+++ server/camcops_server/tasks/phq9.py
@@ -5,7 +5,7 @@ camcops_server/tasks/phq9.py
 
 ===============================================================================
 
-    Copyright (C) 2012-2019 Rudolf Cardinal (rudolf@pobox.com).
+    Copyright (C) 2012-2020 Rudolf Cardinal (rudolf@pobox.com).
 
     This file is part of CamCOPS.